Implement GH1434 usage probes and archive script
