+++ crates/harness-eval/Cargo.toml
@@ -8,5 +8,6 @@ repository.workspace = true
 rust-version.workspace = true
 
 [dependencies]
+harness-core = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
