+++ crates/harness-server/src/scheduler.rs
@@ -112,6 +112,8 @@ impl Scheduler {
             .events
             .persist_rule_scan(&project_root, &violations)
             .await;
+        let probe_report = harness_core::usage_probe::build_probe_report_event()?;
+        state.observability.events.log(&probe_report).await?;
         let report = generate_health_report(&events, &violations);
         tracing::info!(
             grade = ?report.quality.grade,
@@ -200,6 +202,41 @@ mod tests {
         let expected_root = project_root.path().display().to_string();
         assert_eq!(scan.detail.as_deref(), Some(expected_root.as_str()));
 
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn health_tick_persists_usage_probe_report() -> anyhow::Result<()> {
+        let _lock = crate::test_helpers::HOME_LOCK.lock().await;
+        if !crate::test_helpers::db_tests_enabled().await {
+            return Ok(());
+        }
+        let data_dir = tempfile::tempdir()?;
+        let project_root = tempfile::tempdir()?;
+        let state = make_test_state(data_dir.path(), project_root.path()).await?;
+        harness_core::usage_probe::reset_for_tests();
+        harness_core::usage_probe::record_usage(
+            harness_core::usage_probe::UsageProbeSurface::ThreadManager,
+        );
+
+        Scheduler::run_health_tick(&state).await?;
+
+        let events = state
+            .observability
+            .events
+            .query(&EventFilters {
+                hook: Some("probe_report".to_string()),
+                tool: Some("usage_probe".to_string()),
+                ..Default::default()
+            })
+            .await?;
+        assert_eq!(events.len(), 1);
+        let detail = events[0]
+            .detail
+            .as_deref()
+            .ok_or_else(|| anyhow::anyhow!("missing probe report detail"))?;
+        assert!(detail.contains("\"surface\":\"thread_manager\""));
+        assert!(detail.contains("\"count\":1"));
         Ok(())
     }
