+++ crates/harness-server/src/task_runner/mod.rs
@@ -19,22 +19,203 @@ pub use request::{
     fill_missing_repo_from_project, CreateTaskRequest, PersistedRequestSettings, SystemTaskInput,
     MAX_TASK_PRIORITY,
 };
-pub use spawn::{
-    check_awaiting_deps, effective_turn_timeout, prompt_requires_plan, register_pending_task,
-    resolve_canonical_project, spawn_preregistered_task, spawn_task, spawn_task_awaiting_deps,
+use spawn::{
+    check_awaiting_deps as check_awaiting_deps_impl,
+    register_pending_task as register_pending_task_impl,
+    spawn_preregistered_task as spawn_preregistered_task_impl, spawn_task as spawn_task_impl,
+    spawn_task_awaiting_deps as spawn_task_awaiting_deps_impl,
 };
+pub use spawn::{effective_turn_timeout, prompt_requires_plan, resolve_canonical_project};
 pub use state::{
     RecentFailureTask, RoundResult, SchedulerAuthorityState, SchedulerOwner, SchedulerOwnerKind,
     TaskSchedulerState, TaskState, TaskSummary, TaskWorkflowSummary,
 };
-pub use store::{
-    mutate_and_persist, update_status, TaskStore, TaskSummaryFilter, TaskSummaryPageCursor,
-};
+use store::{mutate_and_persist as mutate_and_persist_impl, update_status as update_status_impl};
+pub use store::{TaskStore, TaskSummaryFilter, TaskSummaryPageCursor};
 pub use types::{TaskFailureKind, TaskId, TaskKind, TaskPhase, TaskStatus};
+
+fn record_task_runner_usage() {
+    harness_core::usage_probe::record_usage(
+        harness_core::usage_probe::UsageProbeSurface::TaskRunner,
+    );
+}
+
+pub async fn spawn_task(
+    store: Arc<TaskStore>,
+    agent: Arc<dyn harness_core::agent::CodeAgent>,
+    reviewer: Option<Arc<dyn harness_core::agent::CodeAgent>>,
+    server_config: Arc<harness_core::config::HarnessConfig>,
+    skills: Arc<tokio::sync::RwLock<harness_skills::store::SkillStore>>,
+    events: Arc<harness_observe::event_store::EventStore>,
+    interceptors: impl Into<crate::task_executor::SharedTurnInterceptors>,
+    req: CreateTaskRequest,
+    workspace_mgr: Option<Arc<crate::workspace::WorkspaceManager>>,
+    permit: crate::task_queue::TaskPermit,
+    completion_callback: Option<CompletionCallback>,
+    issue_workflow_store: Option<Arc<harness_workflow::issue_lifecycle::IssueWorkflowStore>>,
+    workflow_runtime_store: Option<Arc<harness_workflow::runtime::WorkflowRuntimeStore>>,
+    allowed_project_roots: Vec<std::path::PathBuf>,
+) -> TaskId {
+    record_task_runner_usage();
+    spawn_task_impl(
+        store,
+        agent,
+        reviewer,
+        server_config,
+        skills,
+        events,
+        interceptors,
+        req,
+        workspace_mgr,
+        permit,
+        completion_callback,
+        issue_workflow_store,
+        workflow_runtime_store,
+        allowed_project_roots,
+    )
+    .await
+}
+
+pub async fn register_pending_task(store: Arc<TaskStore>, req: &CreateTaskRequest) -> TaskId {
+    record_task_runner_usage();
+    register_pending_task_impl(store, req).await
+}
+
+pub async fn spawn_preregistered_task(
+    task_id: TaskId,
+    store: Arc<TaskStore>,
+    agent: Arc<dyn harness_core::agent::CodeAgent>,
+    reviewer: Option<Arc<dyn harness_core::agent::CodeAgent>>,
+    server_config: Arc<harness_core::config::HarnessConfig>,
+    skills: Arc<tokio::sync::RwLock<harness_skills::store::SkillStore>>,
+    events: Arc<harness_observe::event_store::EventStore>,
+    interceptors: impl Into<crate::task_executor::SharedTurnInterceptors>,
+    req: CreateTaskRequest,
+    workspace_mgr: Option<Arc<crate::workspace::WorkspaceManager>>,
+    permit: crate::task_queue::TaskPermit,
+    completion_callback: Option<CompletionCallback>,
+    issue_workflow_store: Option<Arc<harness_workflow::issue_lifecycle::IssueWorkflowStore>>,
+    workflow_runtime_store: Option<Arc<harness_workflow::runtime::WorkflowRuntimeStore>>,
+    allowed_project_roots: Vec<std::path::PathBuf>,
+    group_permit: Option<tokio::sync::OwnedSemaphorePermit>,
+) {
+    record_task_runner_usage();
+    spawn_preregistered_task_impl(
+        task_id,
+        store,
+        agent,
+        reviewer,
+        server_config,
+        skills,
+        events,
+        interceptors,
+        req,
+        workspace_mgr,
+        permit,
+        completion_callback,
+        issue_workflow_store,
+        workflow_runtime_store,
+        allowed_project_roots,
+        group_permit,
+    )
+    .await;
+}
+
+pub async fn spawn_task_awaiting_deps(
+    store: Arc<TaskStore>,
+    req: CreateTaskRequest,
+) -> anyhow::Result<TaskId> {
+    record_task_runner_usage();
+    spawn_task_awaiting_deps_impl(store, req).await
+}
+
+pub async fn check_awaiting_deps(store: &TaskStore) -> (Vec<TaskId>, Vec<TaskId>) {
+    record_task_runner_usage();
+    check_awaiting_deps_impl(store).await
+}
+
+pub async fn update_status(
+    store: &TaskStore,
+    task_id: &TaskId,
+    status: TaskStatus,
+    turn: u32,
+) -> anyhow::Result<()> {
+    record_task_runner_usage();
+    update_status_impl(store, task_id, status, turn).await
+}
+
+pub async fn mutate_and_persist(
+    store: &TaskStore,
+    id: &TaskId,
+    f: impl FnOnce(&mut TaskState),
+) -> anyhow::Result<()> {
+    record_task_runner_usage();
+    mutate_and_persist_impl(store, id, f).await
+}
 
 impl TaskStore {
     /// Return the most recent `limit` failed tasks, newest first.
     pub async fn list_recent_failed(&self, limit: i64) -> anyhow::Result<Vec<RecentFailureTask>> {
+        record_task_runner_usage();
         self.db.list_recent_failed(limit).await
+    }
+}
+
+#[cfg(test)]
+mod usage_probe_tests {
+    use super::*;
+
+    fn task_runner_count() -> u64 {
+        harness_core::usage_probe::snapshot()
+            .into_iter()
+            .find(|entry| entry.surface == "task_runner")
+            .map(|entry| entry.count)
+            .unwrap_or(0)
+    }
+
+    #[tokio::test]
+    async fn task_store_cache_public_entries_record_task_runner_usage() -> anyhow::Result<()> {
+        let _lock = crate::test_helpers::HOME_LOCK.lock().await;
+        if !crate::test_helpers::db_tests_enabled().await {
+            return Ok(());
+        }
+        let dir = tempfile::tempdir()?;
+        let store = TaskStore::open(&dir.path().join("tasks.db")).await?;
+        let task_id = TaskId::from_str("probe-task");
+        let mut task = TaskState::new(task_id.clone());
+        task.project_root = Some(dir.path().to_path_buf());
+        task.repo = Some("owner/repo".to_string());
+        task.external_id = Some("issue-1".to_string());
+        store.insert(&task).await;
+        let mut child = TaskState::new(TaskId::from_str("probe-child"));
+        child.parent_id = Some(task_id.clone());
+        child.project_root = Some(dir.path().to_path_buf());
+        store.insert(&child).await;
+        harness_core::usage_probe::reset_for_tests();
+
+        assert!(store.get(&task_id).is_some());
+        assert_eq!(store.count(), 2);
+        assert_eq!(store.list_all().len(), 2);
+        assert!(store
+            .list_all_statuses_with_terminal()
+            .await?
+            .contains_key(&task_id));
+        assert!(store
+            .list_latest_external_statuses_by_project_and_repo(
+                &dir.path().to_string_lossy(),
+                Some("owner/repo"),
+            )
+            .await?
+            .contains_key("issue-1"));
+        assert!(store.list_siblings(dir.path(), &task_id).is_empty());
+        let since = chrono::Utc::now() - chrono::TimeDelta::days(1);
+        assert_eq!(store.count_done_since(since).await, 0);
+        assert!(store.done_per_project_hour_since(since).await.is_empty());
+        assert_eq!(store.list_children(&task_id).len(), 1);
+        assert!(store.list_recent_failed(5).await?.is_empty());
+        assert!(!store.abort_task(&task_id));
+
+        assert_eq!(task_runner_count(), 11);
+        Ok(())
     }
 }
